requirements.txt
google-generativeai
streamlit
pillow